move tests to same named module folder
